==> 01_FeatureExtractionPart01.ipynb <==
## Initial Setup

from pyspark.sql import SparkSession

#from pyspark.sql import SparkSession
spark = SparkSession.builder.appName("Basics").getOrCreate()

spark

# Add all import commands to this cell. 
from pyspark import SparkConf, SparkContext
from pyspark.sql import SQLContext, types
from pyspark.sql.functions import col, round, regexp_replace, lit, split, to_date, datediff, current_date, months_between
#import sys
#import string
from pyspark.sql.types import StructType, FloatType, ArrayType, StringType, DateType
from pyspark.sql import functions as F



## Read the data

raw_recipes_df = spark.read.csv("s3a://raw-recipes-clean-upgrad/RAW_recipes_cleaned.csv", 
                                 header=True, inferSchema=True
                               )

# Please forward the exact name of data frames and columns as suggested in the code. 
# It will ensure that the assert commands function correctly.  

# Code check cell
# Do not edit cells with assert commands
# If an error is shown after running this cell, please recheck your code.  

assert raw_recipes_df.count() == 231637, "There is a mistake in reading the data."
assert len(raw_recipes_df.columns) == 12, "There is a mistake in reading the data."

## Extract ```nutrition``` values 

raw_recipes_df.select('nutrition').show(5, truncate=False)

# List of nutrition columns

nutrition_column_names = ['calories',
                          'total_fat_PDV',
                          'sugar_PDV',
                          'sodium_PDV',
                          'protein_PDV',
                          'saturated_fat_PDV',
                          'carbohydrates_PDV']

# STEP 1 - string operations to convert nutrition to an array 


raw_recipes_df = (raw_recipes_df
                  .withColumn("nutrition",  regexp_replace("nutrition", "\[", ""))
                  .withColumn("nutrition",  regexp_replace("nutrition", "\]", ""))
                  .withColumn('nutrition',  split(col("nutrition"),",")  ) # add command(s) to modify the string
                  )

#split the text 


raw_recipes_df.select('nutrition').show(5, truncate=False)

# STEP 2 - create new columns with individual nutrition values 

# Also eusure that new columns are float type

raw_recipes_df = (raw_recipes_df
                  .withColumn('calories', raw_recipes_df.nutrition[0].cast('float'))
                  .withColumn('total_fat_PDV', raw_recipes_df.nutrition[1].cast('float'))
                  .withColumn('sugar_PDV', raw_recipes_df.nutrition[2].cast('float'))
                  .withColumn('sodium_PDV', raw_recipes_df.nutrition[3].cast('float'))
                  .withColumn('protein_PDV', raw_recipes_df.nutrition[4].cast('float'))
                  .withColumn('saturated_fat_PDV', raw_recipes_df.nutrition[5].cast('float'))
                  .withColumn('carbohydrates_PDV', raw_recipes_df.nutrition[6].cast('float'))
                 )


# Code check cell
# Do not edit cells with assert commands
# If an error is shown after running this cell, please recheck your code.  

assert raw_recipes_df.schema["carbohydrates_PDV"].dataType == FloatType(), "Recheck your typecasting"
assert raw_recipes_df.collect()[123432][14] == 62.0, "The columns have not been split correctly."
assert raw_recipes_df.collect()[10000][12] == 60.400001525878906, "The columns have not been split correctly."

## Make nutrition-per-100 calorie columns

By converting the nutrition values from absolute to relative terms, we ensure that portion size is not a factor in the analysis. 


Naming convention: Original column name ```total fat (PDV)```, column name after column ```total_fat_per_100_cal```

for nutrition_col in nutrition_column_names:
    if nutrition_col != 'calories':
        nutrition_per_100_cal_col = (nutrition_col
                                 .replace('_PDV','')
                                 +'_per_100_cal')
        # Write code to create nutrition columns per 100 calories 
nutrition_column_names
        
        

from pyspark.sql.functions import udf
#https://stackoverflow.com/questions/48282321/valueerror-cannot-convert-column-into-bool

@udf("double")    
def nutritonPercent(x,y):
    if y <= 0:
        return 0.0
    return ((x / y) * 100)

raw_recipes_df = (raw_recipes_df
                  .withColumn('total_fat_per_100_cal', nutritonPercent(raw_recipes_df.total_fat_PDV, raw_recipes_df.calories))
                  .withColumn('sugar_per_100_cal', nutritonPercent(raw_recipes_df.sugar_PDV, raw_recipes_df.calories))
                  .withColumn('sodium_per_100_cal',  nutritonPercent(raw_recipes_df.sodium_PDV, raw_recipes_df.calories))
                  .withColumn('protein_per_100_cal',  nutritonPercent(raw_recipes_df.protein_PDV, raw_recipes_df.calories))
                  .withColumn('saturated_fat_per_100_cal',  nutritonPercent(raw_recipes_df.saturated_fat_PDV, raw_recipes_df.calories))
                  .withColumn('carbohydrates_per_100_cal',  nutritonPercent(raw_recipes_df.carbohydrates_PDV, raw_recipes_df.calories))
                 )

# total fat check for id 28881
assert raw_recipes_df.filter("id == 28881").select('total_fat_per_100_cal').first()[0] == 0, "total_fat_per_100_cal for recipe 28881 should be 0"

# total fat check for id 112140
assert round(raw_recipes_df.filter("id == 112140").select('total_fat_per_100_cal').first()[0]) == 8, "total_fat_per_100_cal for recipe 112140 should be 8"

## Convert `Tags` to `array` type

raw_recipes_df.select('tags').show(5, truncate=False)

raw_recipes_df.dtypes

raw_recipes_df=(raw_recipes_df
    .withColumn('tags',regexp_replace('tags','\[',''))
    .withColumn('tags',regexp_replace('tags','\]',''))
    .withColumn('tags',  regexp_replace('tags', '\\s', ''))
    .withColumn('tags',  regexp_replace('tags', '\'', ''))
    .withColumn('tags',  split(col("tags"),",")  ) # add command(s) to modify the string
   )

# Code check cell
# Do not edit cells with assert commands
# If an error is shown after running this cell, please recheck your code.  

assert raw_recipes_df.schema["tags"].dataType == ArrayType(StringType(), True), "You have not split the string into an array."
assert raw_recipes_df.collect()[2][5] == ['time-to-make','course', 'preparation', 'main-dish', 'chili', 'crock-pot-slow-cooker', 'dietary', 'equipment', '4-hours-or-less'], "Recheck your string cleaning and splitting operations."

## Join Recipe Data to Review Data

# Reading the second data set. 
# keep this cell unedited

raw_ratings_df = (spark.read.csv("s3a://raw-interactions-upgrad/RAW_interactions_cleaned.csv", 
                                # add arguments
                                 header=True, inferSchema=True
                                )
                  .withColumn("review_date",  F.col("date"))
                  .drop(F.col("date"))
                  )

raw_ratings_df.printSchema()

# Code check cell
# Do not edit cells with assert commands
# If an error is shown after running this cell, please recheck your code.  

assert raw_ratings_df.count() == 1132367, "There is a mistake in reading the data."
assert len(raw_ratings_df.columns) == 5, "There is a mistake in reading the data."

raw_ratings_df.show(5)

interaction_level_df = raw_ratings_df.join(
                                            # code to join raw_recipes to raw_reviews
                                            raw_recipes_df, raw_ratings_df.recipe_id == raw_recipes_df.id, "inner"
                                           )




# Code check cell
# Do not edit cells with assert commands
# If an error is shown after running this cell, please recheck your code.  

list1 = raw_ratings_df.select('recipe_id').collect()
list2 = raw_recipes_df.select('id').collect()
exclusive_set = set(list1)-set(list2)

assert len(exclusive_set) == 0, "There is a mistake in reading one of the two data files."

interaction_level_df.show(5)

interaction_level_df.dtypes

## Recipe Submission Date Based Features

**Steps:**
1. Convert ```submitted``` column to `DateType()`
2. Use review date to derive columns `days_since_submission_on_review_date`,  `months_since_submission_on_review_date`,  `years_since_submission_on_review_date`

interaction_level_df = (interaction_level_df
                        .withColumn('submitted', to_date("submitted"))
                        .withColumn('review_date', to_date("review_date"))
                       )

interaction_level_df.select('submitted', 'review_date').show(5, truncate=False)

interaction_level_df = (interaction_level_df
                        .withColumn('days_since_submission_on_review_date',
                                    # code to calculate days between review and submission
                                    datediff(col("review_date"), col("submitted"))
                                   )
                        .withColumn('months_since_submission_on_review_date',
                                    # code to calculate months between review and submission  
                                    months_between(col("review_date"), col("submitted"))
                                   )
                        .withColumn('years_since_submission_on_review_date',
                                    # code to calculate years between review and submission   
                                    months_between(col("review_date"), col("submitted"))/lit(12)
                                   )
                                   )
                        


interaction_level_df.select('submitted', 'review_date','days_since_submission_on_review_date', 'months_since_submission_on_review_date', 'years_since_submission_on_review_date').show(5, truncate=False)

## Save the data we have created so far in a parquet file. 

interaction_level_df.printSchema()

## Write the raw_recipes_df
## Write data to the S3 bucket which will be your working directory 

interaction_level_df.write.mode('overwrite').parquet( 
                                                    # add code here 
                                                    "s3://sparkml/receiperecommender/interaction_level_df.parquet"
                                                     )
