==> 03_FeatureExtractionPart.ipynb <==
## Initial Setup

from pyspark.sql import SparkSession

from pyspark.sql import SparkSession
spark = SparkSession.builder.appName("Basics").getOrCreate()

spark

# Run this everytime you create a new spark instance. 

spark.sparkContext.install_pypi_package("plotly==5.5.0")
spark.sparkContext.install_pypi_package("pandas==0.25.1")
spark.sparkContext.install_pypi_package("numpy==1.14.5")
spark.sparkContext.install_pypi_package("matplotlib==3.1.1")

# Write all import commands here 
# importing window from pyspark.sql.window
from pyspark.sql.window import Window
from pyspark.sql.functions import col,avg,sum,min,max,row_number, count, when, isnan
from pyspark.sql import functions as F
import sys
from pyspark.ml.feature import StringIndexer

df = sqlContext.createDataFrame( [(1,2,"a"),(3,2,"a"),(1,3,"b"),(2,2,"a"),(2,3,"b")], 
                                 ["time", "value", "class"] )

from pyspark.sql import Window
from pyspark.sql import functions as F

windowval = (Window.partitionBy('class').orderBy('time')
             .rangeBetween(Window.unboundedPreceding, 0))
df_w_cumsum = df.withColumn('cum_sum', F.sum('value').over(windowval))
df_w_cumsum.show()

df_w_cumsum.filter('cum_sum > 3').show()

import matplotlib.pyplot as plt
import numpy as np

xpoints = np.array([1, 8])
ypoints = np.array([3, 10])
print('xpoints are - ', xpoints)

plt.plot(xpoints, ypoints)
plt.show()

## Defining Custom Functions 

def get_quantiles(df, col_name, quantiles_list=[0.01, 0.25, 0.5, 0.75, 0.99]):
    """
    Takes a numerical column and returns column values at requested quantiles

    Inputs 
    Argument 1: Dataframe
    Argument 2: Name of the column
    Argument 3: A list of quantiles you want to find. Default value [0.01, 0.25, 0.5, 0.75, 0.99]

    Output 
    Returns a dictionary with quantiles as keys and column quantile values as values 
    """
    # Get min, max and quantile values for given column
    zero = df.approxQuantile(col_name,[0.0] , 0.2)[0]
    quant_vals = df.approxQuantile(col_name, quantiles_list, 0.2)
    one = df.approxQuantile(col_name,[1.0] , 0.2)[0]

    # Store min, quantiles and max in output dict, sequentially
    quantiles_dict = {0.0:int(zero)}
    for i in range(0,len(quantiles_list)):
        quantiles_dict[quantiles_list[i]]=quant_vals[i]
    quantiles_dict[1.0]=int(one)

    return(quantiles_dict)

def add_OHE_columns (df, n_name_list):
    """
    Given a list of tags, creates one hot encoded columns for each tag. 
  
    Input
    Argument 1: Dataframe in which the function will add the new columns
    Argument 2: list of tags
  
    Output
    Prints the names of columns that have been added 
    Returns the modified dataframe 
    """
    # write code to create encoded columns 
    for name in n_name_list:
        opt_col = name+'_numeric'
        indexer = StringIndexer(inputCol=name, outputCol=opt_col)
        df = indexer.fit(df).transform(df)
        oht_opt_col = name+'_onehot'
        encoder = OneHotEncoder(inputCols=['opt_col'], outputCols=[oht_opt_col])
        df = encoder.fit(df).transform(df)
        df = df.select('*', vector_to_array('oht_opt_col').alias('col_onehot'))        
        print ("added column: has_tag_"+name)

    return df

df = add_OHE_columns(interaction_level_df, top_frequent_tags_list)
df.show()

#sample reference - https://www.skytowner.com/explore/one_hot_encoding_in_pyspark

rows = [['Alex','B'], ['Bob','A'], ['Cathy','B'], ['Dave','C'], ['Eric','D']]
df = spark.createDataFrame(rows, ['name','class'])
df.show()

from pyspark.ml.feature import StringIndexer
indexer = StringIndexer(inputCol='class', outputCol='class_numeric')
df_indexed = indexer.fit(df).transform(df)
df_indexed.show()

from pyspark.ml.feature import OneHotEncoder
encoder = OneHotEncoder(inputCols=['class_numeric'], outputCols=["oht_opt_col"])
df_onehot = encoder.fit(df_indexed).transform(df_indexed)
df_onehot.show()

from pyspark.ml.functions import vector_to_array
df_col_onehot = df_onehot.select('*', vector_to_array('oht_opt_col').alias('col_onehot'))
df_col_onehot.show()

## Read the data

interaction_level_df = spark.read.parquet(
                                            # add code here 
                                            "s3://sparkml/EDA02/interaction_level_df.parquet"
                                        )

## Adding user level average features

1. user_avg_rating
2. user_n_ratings
3. user_avg_years_betwn_review_and_submission
4. user_avg_prep_time_recipes_reviewed
5. user_avg_n_steps_recipes_reviewed
6. user_avg_n_ingredients_recipes_reviewed

# creating a window partition of dataframe
# for - 1. user_avg_rating
userAvgRatingAgg  = Window.partitionBy("user_id")

# calculate average() of ratings per user
interaction_level_df = interaction_level_df.withColumn("user_avg_rating",
               avg(col("rating")).over(userAvgRatingAgg))

# for - 2. user_n_ratings
interaction_level_df = interaction_level_df.withColumn("user_n_ratings",
               count(col("rating")).over(userAvgRatingAgg))

# for - 3. user_avg_years_betwn_review_and_submission
interaction_level_df = interaction_level_df.withColumn("user_avg_years_betwn_review_and_submission",
               avg(col("years_since_submission_on_review_date")).over(userAvgRatingAgg))

# for - 4. user_avg_prep_time_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_prep_time_recipes_reviewed",
               avg(col("minutes")).over(userAvgRatingAgg))

# for 5. user_avg_n_steps_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_n_steps_recipes_reviewed",
               avg(col("n_steps")).over(userAvgRatingAgg))

# for 6. user_avg_n_ingredients_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_n_ingredients_recipes_reviewed",
               avg(col("n_ingredients")).over(userAvgRatingAgg))


Add columns 

7. user_avg_calories

8. user_avg_total_fat_per_100_cal_recipes_reviewed

9. user_avg_sugar_per_100_cal_recipes_reviewed

10. user_avg_sodium_per_100_cal_recipes_reviewed

11. user_avg_protein_per_100_cal_recipes_reviewed

12. user_avg_saturated_fat_per_100_cal_recipes_reviewed

13. user_avg_carbohydrates_per_100_cal_recipes_reviewed

# for - 7. user_avg_calories
interaction_level_df = interaction_level_df.withColumn("user_avg_calories",
               avg(col("calories")).over(userAvgRatingAgg))


#for 8. user_avg_total_fat_per_100_cal_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_total_fat_per_100_cal_recipes_reviewed",
               avg(col("total_fat_per_100_cal")).over(userAvgRatingAgg))
#round(interaction_level_df.filter('user_id == 233044').select('user_avg_total_fat_per_100_cal_recipes_reviewed').first()[0])

#for - 9. user_avg_sugar_per_100_cal_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_sugar_per_100_cal_recipes_reviewed",
               avg(col("sugar_per_100_cal")).over(userAvgRatingAgg))

#for 10. user_avg_sodium_per_100_cal_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_sodium_per_100_cal_recipes_reviewed",
               avg(col("sodium_per_100_cal")).over(userAvgRatingAgg))

#for 11. user_avg_protein_per_100_cal_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_protein_per_100_cal_recipes_reviewed",
               avg(col("protein_per_100_cal")).over(userAvgRatingAgg))

#for 12. user_avg_saturated_fat_per_100_cal_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_saturated_fat_per_100_cal_recipes_reviewed",
               avg(col("saturated_fat_per_100_cal")).over(userAvgRatingAgg))

#for 13. user_avg_carbohydrates_per_100_cal_recipes_reviewed
interaction_level_df = interaction_level_df.withColumn("user_avg_carbohydrates_per_100_cal_recipes_reviewed",
               avg(col("carbohydrates_per_100_cal")).over(userAvgRatingAgg))

# Code check cell
# Do not edit cells with assert commands
# If an error is shown after running this cell, please recheck your code. 

assert(round(interaction_level_df.filter('user_id == 601529').select('user_avg_rating').first()[0], 2) == 4.22)
assert(interaction_level_df.filter('user_id == 601529').select('user_n_ratings').first()[0] == 27)
assert(round(interaction_level_df.filter('user_id == 601529').select('user_avg_years_betwn_review_and_submission').first()[0], 2) == 3.51)
assert(interaction_level_df.filter('user_id == 233044').select('user_avg_prep_time_recipes_reviewed').first()[0] == 50.3)
assert(interaction_level_df.filter('user_id == 233044').select('user_avg_n_steps_recipes_reviewed').first()[0] == 8.8)
assert(interaction_level_df.filter('user_id == 233044').select('user_avg_n_ingredients_recipes_reviewed').first()[0] == 8.2)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_total_fat_per_100_cal_recipes_reviewed').first()[0]) == 6)

**More Features:**

high_ratings = 5 rating

- `user_avg_years_betwn_review_and_submission_high_ratings`
- `user_avg_prep_time_recipes_reviewed_high_ratings`
- `user_avg_n_steps_recipes_reviewed_high_ratings`
- `user_avg_n_ingredients_recipes_reviewed_high_ratings`

#from pyspark.sql.functions import when
# Create indicator columns for highest ratings 
interaction_level_df = interaction_level_df.withColumn("ind_5_rating",
               when(col("rating") ==5, "1").otherwise("0"))



#user_n_5_ratings
interaction_level_df = (
                        interaction_level_df
                        .withColumn("user_n_5_ratings",
                                    when(col("rating") ==5, count(col("rating")).over(userAvgRatingAgg)).otherwise(None))
                       )

interaction_level_df.filter('user_id == 601529').select('user_id',  'rating', 'user_n_5_ratings').show(5)

# Add columns mentioned above 
# hint explore window functions in pyspark.sql
#user_avg_years_betwn_review_and_submission_5_ratings: double (nullable = true)
interaction_level_df = (
                        interaction_level_df
                        .withColumn("user_avg_years_betwn_review_and_submission_5_ratings",
                        when(col("rating") ==5, avg(col("years_since_submission_on_review_date")).over(userAvgRatingAgg))
                        .otherwise(None))
                       )

# `user_avg_prep_time_recipes_reviewed_high_ratings`
#user_avg_prep_time_recipes_reviewed_5_ratings: double (nullable = true)
interaction_level_df = (
                        interaction_level_df
                        .withColumn("user_avg_prep_time_recipes_reviewed_5_ratings",
                        when(col("rating") ==5, avg(col("minutes")).over(userAvgRatingAgg))
                        .otherwise(None))
                       )

# user_avg_n_steps_recipes_reviewed: double (nullable = true)
interaction_level_df = (
                        interaction_level_df
                        .withColumn("user_avg_n_steps_recipes_reviewed_5_ratings",
                        when(col("rating") ==5, avg(col("n_steps")).over(userAvgRatingAgg))
                        .otherwise(None))
                       )

# user_avg_n_ingredients_recipes_reviewed: double (nullable = true)
interaction_level_df = (
                        interaction_level_df
                        .withColumn("user_avg_n_ingredients_recipes_reviewed_5_ratings",
                        when(col("rating") ==5, avg(col("n_ingredients")).over(userAvgRatingAgg))
                        .otherwise(None))
                       )



#calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("calories_5_ratings",
                                   when(col("rating") ==5, col("calories")).otherwise(None)
                                  )
                       )

interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_calories_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("calories")).over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )
#fat per 100 calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("total_fat_per_100_cal_5_ratings",
                                   when(col("rating") ==5, col("total_fat_per_100_cal")).otherwise(None)
                                  )
                       )
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_total_fat_per_100_cal_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("total_fat_per_100_cal"))
                                        .over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )

# sugar per 100 calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("sugar_per_100_cal_5_ratings",
                                   when(col("rating") ==5, col("sugar_per_100_cal")).otherwise(None)
                                  )
                       )
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_sugar_per_100_cal_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("sugar_per_100_cal"))
                                        .over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )

# sodium per 100 calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("sodium_per_100_cal_5_ratings",
                                   when(col("rating") ==5, col("sodium_per_100_cal")).otherwise(None)
                                  )
                       )
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_sodium_per_100_cal_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("sodium_per_100_cal"))
                                        .over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )

# protein per 100 calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("protein_per_100_cal_5_ratings",
                                   when(col("rating") ==5, col("protein_per_100_cal")).otherwise(None)
                                  )
                       )
    
    
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_protein_per_100_cal_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("protein_per_100_cal"))
                                        .over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )

# saturated_fat per 100 calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("saturated_fat_per_100_cal_5_ratings",
                                   when(col("rating") ==5, col("saturated_fat_per_100_cal")).otherwise(None)
                                  )
                       )
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_saturated_fat_per_100_cal_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("saturated_fat_per_100_cal"))
                                        .over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )

# carbohydrates per 100 calories columns
interaction_level_df = (
                        interaction_level_df.
                        withColumn("carbohydrates_per_100_cal_5_ratings",
                                   when(col("rating") ==5, col("carbohydrates_per_100_cal")).otherwise(None)
                                  )
                       )
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_carbohydrates_per_100_cal_recipes_reviewed_5_ratings",
                                   when(col("rating") ==5, avg(col("carbohydrates_per_100_cal"))
                                        .over(userAvgRatingAgg)).otherwise(None)
                                  )
                       )

interaction_level_df.filter('user_id == 601529').select('user_id', 'calories', 'rating', 'calories_5_ratings').show(5)

# it is possible that you might have added null values in the dataframe as a result of this exercise. 
# Identify the null values and deal with them 
#from pyspark.sql.functions import col,isnan, when, count
NewColumns = ["user_n_5_ratings", 
              "user_avg_years_betwn_review_and_submission_5_ratings",
              "user_avg_prep_time_recipes_reviewed_5_ratings",
              "user_avg_n_steps_recipes_reviewed_5_ratings",
              "user_avg_n_ingredients_recipes_reviewed_5_ratings",
              "calories_5_ratings",
              "user_avg_calories_recipes_reviewed_5_ratings",
              "total_fat_per_100_cal_5_ratings",
              "user_avg_total_fat_per_100_cal_recipes_reviewed_5_ratings",
              "sugar_per_100_cal_5_ratings",
              "user_avg_sugar_per_100_cal_recipes_reviewed_5_ratings",
              "sodium_per_100_cal_5_ratings",
              "user_avg_sodium_per_100_cal_recipes_reviewed_5_ratings",
              "protein_per_100_cal_5_ratings",
              "user_avg_protein_per_100_cal_recipes_reviewed_5_ratings",
              "saturated_fat_per_100_cal_5_ratings",
              "user_avg_saturated_fat_per_100_cal_recipes_reviewed_5_ratings",
              "carbohydrates_per_100_cal_5_ratings",
              "user_avg_carbohydrates_per_100_cal_recipes_reviewed_5_ratings"
             ]
interaction_level_df.select([count(when(isnan(c) | col(c).isNull(), c)).alias(c) for c in NewColumns]
   ).count()


#user_avg_5_ratings
interaction_level_df = (
                        interaction_level_df.
                        withColumn("user_avg_5_ratings",
                                   when(col("rating") ==5, avg(col("rating"))
                                        .over(userAvgRatingAgg))
                                  )
                       )

# Check - All rows with ratings should have non-null values in corresponding user_avg_5_ratings columns
#interaction_level_df.filter(col("user_avg_5_ratings").isNull()).select('rating','user_avg_5_ratings').show(5)
interaction_level_df.filter(col("user_avg_5_ratings").isNull()).count()


interaction_level_df.filter('user_id == 233044').select('user_id', 'rating', 'user_avg_5_ratings').show()

interaction_level_df.filter('user_id == 233044').select('user_n_5_ratings').first()[0]

interaction_level_df.groupby("user_n_5_ratings").count().show(30)

# taking a backup 
interaction_level_df_backup = interaction_level_df

interaction_level_df = interaction_level_df.filter(interaction_level_df.user_n_5_ratings.isNotNull())

interaction_level_df.count()

# Check values for a given user id

assert(interaction_level_df.filter('user_id == 233044').select('user_n_5_ratings').first()[0] == 7 )
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_years_betwn_review_and_submission_5_ratings').first()[0], 2) == 2.24 )
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_prep_time_recipes_reviewed_5_ratings').first()[0]) == 46)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_n_steps_recipes_reviewed_5_ratings').first()[0], 2) == 7.29)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_n_ingredients_recipes_reviewed_5_ratings').first()[0], 2) == 6.86)

interaction_level_df.filter('user_id == 233044').select('user_n_5_ratings').show()

interaction_level_df.filter('user_id == 233044').select('user_avg_years_betwn_review_and_submission_5_ratings').show()

# years_since_submission_on_review_date_5_ratings
interaction_level_df = (
                        interaction_level_df.
                        withColumn("years_since_submission_on_review_date_5_ratings",
                                   when(col("rating") ==5, col("years_since_submission_on_review_date")).otherwise(None)
                                  )
                       )
#minutes_5_ratings
interaction_level_df = (
                        interaction_level_df.
                        withColumn("minutes_5_ratings",
                                   when(col("rating") ==5, col("minutes")).otherwise(None)
                                  )
                       )

#n_steps_5_ratings
interaction_level_df = (
                        interaction_level_df.
                        withColumn("n_steps_5_ratings",
                                   when(col("rating") ==5, col("n_steps")).otherwise(None)
                                  )
                       )

#n_ingredients_5_ratings
interaction_level_df = (
                        interaction_level_df.
                        withColumn("n_ingredients_5_ratings",
                                   when(col("rating") ==5, col("n_ingredients")).otherwise(None)
                                  )
                       )

interaction_level_df.printSchema()

## Tags level EDA

interaction_tag_level_df = interaction_level_df.withColumn('individual_tag',F.explode('tags'))

interaction_tag_level_df.count()

tags_ratings_summary = (interaction_tag_level_df
                        .groupBy('individual_tag').agg(F.avg('rating').alias('avg_user_rating'),
#                                                      F.max('rating').alias('max_user_rating'),
#                                                      F.min('rating').alias('min_user_rating'),
                                                       F.count('rating').alias('n_user_ratings'),
                                                       F.countDistinct('id').alias('n_recipes')))

interactions, recipes  =  interaction_level_df.count(), interaction_level_df.agg(F.countDistinct('id')).first()[0]

tags_ratings_summary = (tags_ratings_summary.withColumn("in_percent_recipies", F.col ("n_recipes")/F.lit(recipes))
                                            .withColumn("in_percent_interactions", F.col ("n_user_ratings")/F.lit(interactions)))

#### 1. Top ```n``` most rated tags

tags_ratings_summary.sort(tags_ratings_summary.n_user_ratings.desc()).show(20)

Drop tags present in majority of recipes more than 75% rows 

tags_ratings_summary_1 = tags_ratings_summary.filter('in_percent_interactions <= 0.75')

tags_ratings_summary_1.count()

Percentlile95th = 527 * 0.95

Percentlile95th

tags_ratings_summary_1.sort(tags_ratings_summary_1.in_percent_interactions.desc()).show(29)

tags_ratings_summary_1.createOrReplaceTempView("S_tag_sql")

tags_ratings_summary_temp = spark.sql("""SELECT *
FROM S_tag_sql  
order by S_tag_sql.in_percent_interactions  desc limit 29""")

tags_ratings_summary_temp.show(30)

# print the quantiles of in_percent_interactions  
get_quantiles(df = tags_ratings_summary_1, # add code here , 
              col_name = "in_percent_interactions", # add code here , 
              quantiles_list = [0.01,0.25,0.5, 0.75,0.8,0.85,0.9,0.95, 0.99])

# Actual output and expected output is not matching -- pending validation -- Jagadish 9/18/2022

# keep tags appearing in the top 5 percentile 
top_most_frequent_tags = tags_ratings_summary_temp

top_most_frequent_tags.count()

# extract the list of top frequent tags
top_frequent_tags_list =top_most_frequent_tags.select('individual_tag').rdd.flatMap(lambda x: x).collect()

top_frequent_tags_list

interaction_level_df = add_OHE_columns (
 # add code here 
    interaction_level_df, top_frequent_tags_list
)

#### 2.  Bottom ```n``` least rated tags

The above tags are present in 1 recipe in over two hundred thousand. The features we create based on these tags will not teach the model new information. If these tags were one hot encoded, the entire column would be filled with zeros, and only a few rows will have 1s. One hot encoding of these tags is not a good idea. If you come up with an encoding that captures the rarity of these tags, only then can you add these tags to the analysis.

#### 3. Top ```n``` rated tags 

Top rated tags have low number of ratings.

get_quantiles (df = # add code here, 
               col_name = # add code here, 
               quantiles_list = [0.01, 0.05, 0.1, 0.15, 0.2, 0.25, 0.5, 0.75, 0.99])

# Drop the tags with less than 100 ratings 
tags_ratings_summary =

get_quantiles(df = # add code here, 
              col_name = # add code here,
              quantiles_list = [0.01,0.25,0.5, 0.75,0.8,0.85,0.9,0.95, 0.99])

# keep tags above 95 percentile
top_rated_tags_df = 

top_rated_tags_df.count()

top_rated_tags_list = # collect the list of top rated tags 

Check if any of the current tags have been added earlier

set(top_frequent_tags_list) & set(top_rated_tags_list)

all_added_columns_set = set(top_frequent_tags_list).union(set(top_rated_tags_list))

interaction_level_df = add_OHE_columns ( # add code here )

#### 3. Bottom ```n``` rated tags 

Check if any of the current tags have been added earlier

all_added_columns_set & set(bottom_rated_tags_list)

## Final DataFrame

len(interaction_level_df.columns)

interaction_level_df.write.mode('overwrite').parquet(# add code here)
